Fix transform origin on letter scale
Before, the origin wasn't being applied. This meant that when the letter
zoomed, the right side of the arch would exit the screen after the left
side did, which created an unsatisfactory impression of asymmetry. The
delay between the left side of the arch exiting the screen and the right
was too small to be intentional, so it just looked bad.